Move recycle-ics to new hierarchy
