+++ recycle-ics/recycle-ics.cabal
@@ -55,9 +55,9 @@ library
   ghc-options:
     -Wall
   exposed-modules:
-    Recycle
-    Recycle.ICalendar
-    Recycle.Opts
+    Recycle.Ics
+    Recycle.Ics.ICalendar
+    Recycle.Ics.Opts
   other-modules:
     Paths_recycle_ics
     Servant.API.QueryParamForm
